Correción  del formulario
