--- src/view/cli-prov/componentes/style_Form.css
@@ -1,72 +1,0 @@
-.ContenedorPpal{
-    position: relative;
-    left: 70px;
-    width: 999px;
-    height: 600px;
-    margin-top: 50px;
-    background: linear-gradient(180deg, #77B9F2 37.5%, rgba(255, 255, 255, 0) 100%);
-    box-shadow: inset 0px 4px 4px rgba(0, 0, 0, 0.25);
-    border-radius: 36px;
-}
-
-.foco{
-    position: absolute;
-    left: 45%;
-    right: 30%;
-    top: 15.37%;
-    bottom: 2.81%;
-    background: #BBDEEF;
-    border-radius: 23px;
-}
-
-.form-group{
-    position: sticky;
-    margin-left: 65px;
-    
-}
-
-.guardar{
-    height: 0px;
-    margin-left: 10px;
-    font-weight: 600px;  
-}
-
-.cancelar{
-    margin-left: 120px;
-    font-weight: 600px;  
-}
-
-.form-icons{
-    margin-top: 5%;
-    margin-left: 30%;
-    width: 40%;
-    height: 40%;
-}
-
-.titulo-form{
-    margin-left: 18%;
-    font-weight: bold;
-    size: 12%;
-}
-
-.form-control{
-    box-shadow:cadetblue;
-    font-size: smaller;
-    max-height: 60%;
-    min-height: 20%;
-    
-    
-}
-
-.span{
-    font-size: 10px;
-    font-weight: bold;
-    font-style: italic;
-    height: 5px;
-}
-
-
-.row{
-    height: 50px;
-}
-
