+++ src/view/cli-prov/formulario/style_Form.css
@@ -1,0 +1,87 @@
+.ContenedorPpal{
+    position: relative;
+    left: 70px;
+    width: 999px;
+    height: 650px;
+    margin-top: 50px;
+    background: linear-gradient(180deg, #77B9F2 37.5%, rgba(255, 255, 255, 0) 100%);
+    box-shadow: inset 0px 4px 4px rgba(0, 0, 0, 0.25);
+    border-radius: 36px;
+}
+
+.foco{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    position: relative;
+    top: 50%;
+    width: 30%;
+    height: 85vh;
+    min-width: 250px;
+    margin: auto;
+    background: #BBDEEF;
+    border-radius: 23px;
+}
+
+.cliente{
+   display: flex;
+   flex-direction: column;
+   justify-content: center;
+   width: 100%;
+   align-items: center;
+}
+
+.imagen{
+    width: 130px;
+    height: 130px;
+}
+
+.form-group{
+   display: flex;
+   flex-direction: column;
+   justify-content: center;
+   align-items: center;
+   width: 100%;
+}
+
+.guardar{
+    height: 0px;
+    margin-left: 10px;
+    font-weight: 600px;  
+}
+
+.cancelar{
+    margin-left: 120px;
+    font-weight: 600px;  
+}
+
+
+.titulo-form{
+    margin-top: 3%;
+    margin-bottom: 12%;
+    font-weight: bold;
+    size: 12%;
+}
+
+
+
+.span{
+    font-size: 10px;
+    font-weight: bold;
+    font-style: italic;
+    height: 5px;
+}
+
+
+.row{
+    height: 50px;
+    width: 70%;
+    display: flex;
+    justify-content: center;
+}
+
+.botones{
+    display: flex;
+    width: 60%;
+    justify-content: space-between;
+}
